InsightGrid/requirements.txt
Django>=6.0
djangorestframework>=3.14
django-filter>=24.0
djangorestframework-simplejwt>=5.3
psycopg2-binary>=2.9
django-cors-headers>=4.3
django-extensions>=3.2
